+++ psi_cardinality/python/tests/benchmarks.py
@@ -3,33 +3,34 @@ import sys
 import psi_cardinality.python as psi
 
 
-def test_sanity():
-    s = psi.server.CreateWithNewKey()
-    assert s != None
-
-    key = s.GetPrivateKeyBytes()
-    assert key != None
-
-    other = psi.server.CreateFromKey(key)
-    newkey = other.GetPrivateKeyBytes()
-
-    assert key == newkey
+def helper_client_create_request(cnt):
+    c = psi.client.Create()
+    inputs = ["Element " + str(i) for i in range(cnt)]
+    req = c.CreateRequest(inputs)
 
 
-def test_server_client():
+@pytest.mark.parametrize("cnt", [1, 10, 100, 1000, 10000])
+def test_client_create_request(cnt, benchmark):
+    benchmark(helper_client_create_request, cnt)
+
+
+def helper_client_process_response(cnt):
     c = psi.client.Create()
     s = psi.server.CreateWithNewKey()
 
-    client_items = ["Element " + str(i) for i in range(1000)]
-    server_items = ["Element " + str(2 * i) for i in range(10000)]
+    fpr = 1.0 / 1000000
+    inputs = ["Element " + str(i) for i in range(cnt)]
+    req = c.CreateRequest(inputs)
 
-    setup = s.CreateSetupMessage(0.01, len(client_items), server_items)
-    request = c.CreateRequest(client_items)
+    setup = s.CreateSetupMessage(fpr, len(inputs), inputs)
+    request = c.CreateRequest(inputs)
     resp = s.ProcessRequest(request)
     intersection = c.ProcessResponse(setup, resp)
 
-    assert intersection >= (len(client_items) / 2.0)
-    assert intersection <= (1.1 * len(client_items) / 2.0)
+
+@pytest.mark.parametrize("cnt", [1, 10, 100, 1000, 10000])
+def test_client_process_response(cnt, benchmark):
+    benchmark(helper_client_process_response, cnt)
 
 
 def helper_server_setup(cnt, fpr):
