+++ psi_cardinality/python/tests/tests.py
@@ -1,5 +1,6 @@
 import pytest
 import sys
+import re
 import psi_cardinality.python as psi
 
 
@@ -24,34 +25,38 @@ def test_client_server():
     assert intersection <= (1.1 * len(client_items) / 2.0)
 
 
-def helper_client_create_request(cnt):
-    c = psi.client.Create()
-    inputs = ["Element " + str(i) for i in range(cnt)]
-    req = c.CreateRequest(inputs)
+def test_version():
+    version = psi.__version__
+    assert re.match(r"[0-9]+[.][0-9]+[.][0-9]+(-[A-Za-z0-9]+)?", version)
 
 
-@pytest.mark.parametrize("cnt", [1, 10, 100, 1000, 10000])
-def test_client_create_request(cnt, benchmark):
-    benchmark(helper_client_create_request, cnt)
+def test_sanity():
+    s = psi.server.CreateWithNewKey()
+    assert s != None
+
+    key = s.GetPrivateKeyBytes()
+    assert key != None
+
+    other = psi.server.CreateFromKey(key)
+    newkey = other.GetPrivateKeyBytes()
+
+    assert key == newkey
 
 
-def helper_client_process_response(cnt):
+def test_server_client():
     c = psi.client.Create()
     s = psi.server.CreateWithNewKey()
 
-    fpr = 1.0 / 1000000
-    inputs = ["Element " + str(i) for i in range(cnt)]
-    req = c.CreateRequest(inputs)
+    client_items = ["Element " + str(i) for i in range(1000)]
+    server_items = ["Element " + str(2 * i) for i in range(10000)]
 
-    setup = s.CreateSetupMessage(fpr, len(inputs), inputs)
-    request = c.CreateRequest(inputs)
+    setup = s.CreateSetupMessage(0.01, len(client_items), server_items)
+    request = c.CreateRequest(client_items)
     resp = s.ProcessRequest(request)
     intersection = c.ProcessResponse(setup, resp)
 
-
-@pytest.mark.parametrize("cnt", [1, 10, 100, 1000, 10000])
-def test_client_process_response(cnt, benchmark):
-    benchmark(helper_client_process_response, cnt)
+    assert intersection >= (len(client_items) / 2.0)
+    assert intersection <= (1.1 * len(client_items) / 2.0)
 
 
 if __name__ == "__main__":
